Add the ability to infer additional RuntimeGroups from RIDs
During source build it's desirable to add RIDs to the runtime graph
without actually. Making a change to sources.

For example: the tar-ball is produced and then built on some new distro.

The added RID should follow the compatibility rules for existing RIDs
so we parse it and try to find the most appropriate RuntimeGroup to add
it to.